+++ R/get_metadata.R
@@ -102,9 +102,10 @@ lr_get_metadata <- function(where = getwd(), ext = "ProcSpec", sep = NULL,
     return(NULL)
   } else if (length(whichfailed) > 0) {
 
-    warning("Could not import one or more files:\n",
-            paste0(files[whichfailed], "\n"),
-            call. = FALSE
+    warning(
+      "Could not import one or more files:\n",
+      paste0(files[whichfailed], collapse = "\n"),
+      call. = FALSE
     )
 
     specnames <- specnames[-whichfailed]
@@ -121,9 +122,7 @@ lr_get_metadata <- function(where = getwd(), ext = "ProcSpec", sep = NULL,
     paste(c("white", "dark", "sample"), "boxcar", sep = "_")
   )
 
-  res[, c(6,7,8,9,10,11,12,13,14)] <- vapply(res[, c(6,7,8,9,10,11,12,13,14)],
-                                             as.numeric,
-                                             numeric(nrow(res)))
+  res[, 6:14] <- vapply(res[, 6:14], as.numeric, numeric(nrow(res)))
   res$datetime <- as.POSIXct(res$datetime, tz = "UTC")
 
   return(res)