+++ R/parse_oceanoptics_converted.R
@@ -66,7 +66,7 @@ lr_parse_jaz <- function(filename) {
   }
 
   if (tz == "") {
-    tz = "UTC"
+    tz <- "UTC"
   }
 
   # OceanOptics files use locale-dependent date formats but it looks like they
@@ -112,29 +112,35 @@ lr_parse_jaz <- function(filename) {
 
   # SPECTRA
 
-  data_start <- grep("^(>>>>>Begin (Processed )?Spectral Data<<<<<|>>>>> Comienza Data<<<<< Espectral Procesado Del EL)$", content)
-  data_end <- grep("^(>>>>>End (Processed )?Spectral Data<<<<<|>>>>> Data<<<<< Espectral Procesado Extremo)$", content)
+  data_start <- grep(
+    "^(>>>>>Begin (Processed )?Spectral Data<<<<<|>>>>> Comienza Data<<<<< Espectral Procesado Del EL)$",
+    content
+  )
+  data_end <- grep(
+    "^(>>>>>End (Processed )?Spectral Data<<<<<|>>>>> Data<<<<< Espectral Procesado Extremo)$",
+    content
+  )
 
   # Some files are missing the ending "tag". Let's then assume that data go to
   # the end of file.
-  if (length(data_end)==0) {
+  if (length(data_end) == 0) {
     data_end <- length(content)
   }
 
   # Some files have an extra header for the data, some don't...
   # If they do, it looks like this header will always start with W
-  has_header <- startsWith(content[data_start+1], "W")
+  has_header <- startsWith(content[data_start + 1], "W")
 
-  data <- content[seq(ifelse(has_header, data_start+2, data_start+1),
-                      data_end-1)]
+  data <- content[seq(ifelse(has_header, data_start + 2, data_start + 1),
+                      data_end - 1)]
 
   # Depending on the user locale, some files might use ',' as a decimal sep
-  data <- gsub(",", ".", data)
+  data <- gsub(",", ".", data, fixed = TRUE)
 
-  data <- do.call(rbind, strsplit(data, "\t"))
+  data <- do.call(rbind, strsplit(data, "\t", fixed = TRUE))
 
   if (has_header) {
-    colnames(data) <- strsplit(content[data_start+1], "\t")[[1]]
+    colnames(data) <- strsplit(content[data_start + 1], "\t", fixed = TRUE)[[1]]
 
 
   } else {