+++ tests/testthat/test_dispatch.R
@@ -41,8 +41,8 @@ test_that("Fallback", {
   )
 
   expect_identical(
-    lr_parse_spc(test.file("compare/CRAIC/CRAIC.spc")),
-    dispatch_parser(test.file("compare/CRAIC/CRAIC.spc"))
+    lr_parse_spc(test.file("compare", "CRAIC", "CRAIC.spc")),
+    dispatch_parser(test.file("compare", "CRAIC", "CRAIC.spc"))
   )
 
   expect_identical(
@@ -56,7 +56,7 @@ test_that("Similar output for all parsers", {
 
   files <- list.files(test.file(),
                       recursive = TRUE, include.dirs = TRUE)
-  files <- files[!tools::file_ext(files) %in% c("", "fail", "DRK", "REF")]
+  files <- files[!tools::file_ext(files) %in% c("", "fail", "DRK", "REF", "Raw8", "IRR8")]
 
   lapply(files, function(file) {
     res <- expect_silent(dispatch_parser(test.file(file), sep = ",", specnum = 1))